apps/requirements.txt
asgiref==3.8.1
Django==4.2
sqlparse==0.5.3
tzdata==2025.1
django-ckeditor==6.7.2
django-cors-headers==4.4.0
django-jazzmin==3.0.1
django-js-asset==2.2.0
djangorestframework==3.15.2
django-cors-headers
drf-yasg==1.21.10
packaging==24.2
pillow==10.4.0
psycopg2-binary==2.9.10
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing-extensions==4.12.2
uritemplate==4.1.1
